refactor: split upload registry from receiver
